third_party/WebKit/LayoutTests/html/tabular_data/td_colspan.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("HTMLTableCellElement colspan attribute test");

function colspanAttributeEffect(value)
{
    var element = document.createElement("td");
    element.setAttribute("colSpan", value);
    return element.colSpan;
}

shouldBe('document.createElement("td").colSpan', '1');

shouldBe('colspanAttributeEffect("")', '1');

shouldBe('colspanAttributeEffect("1")', '1');
shouldBe('colspanAttributeEffect("2")', '2');
shouldBe('colspanAttributeEffect("10")', '10');

shouldBe('colspanAttributeEffect("0")', '1');

shouldBe('colspanAttributeEffect("-1")', '1');

shouldBe('colspanAttributeEffect("1x")', '1');
shouldBe('colspanAttributeEffect("1.")', '1');
shouldBe('colspanAttributeEffect("1.9")', '1');
shouldBe('colspanAttributeEffect("2x")', '2');
shouldBe('colspanAttributeEffect("2.")', '2');
shouldBe('colspanAttributeEffect("2.9")', '2');

shouldBe('colspanAttributeEffect("a")', '1');
shouldBe('colspanAttributeEffect("\v7")', '1');
shouldBe('colspanAttributeEffect("  7")', '7');

var arabicIndicDigitOne = String.fromCharCode(0x661);
shouldBe('colspanAttributeEffect(arabicIndicDigitOne)', '1');
shouldBe('colspanAttributeEffect("2" + arabicIndicDigitOne)', '2');

shouldBe('colspanAttributeEffect("2147483647")', '1000');
shouldBe('colspanAttributeEffect("4294967295")', '1000');
shouldBe('colspanAttributeEffect("4294967296")', '1000');
</script>
</body>
</html>
