third_party/blink/web_tests/html/tabular_data/td_colspan-expected.txt

HTMLTableCellElement colspan attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement("td").colSpan is 1
PASS colspanAttributeEffect("") is 1
PASS colspanAttributeEffect("1") is 1
PASS colspanAttributeEffect("2") is 2
PASS colspanAttributeEffect("10") is 10
PASS colspanAttributeEffect("0") is 1
PASS colspanAttributeEffect("-1") is 1
PASS colspanAttributeEffect("1x") is 1
PASS colspanAttributeEffect("1.") is 1
PASS colspanAttributeEffect("1.9") is 1
PASS colspanAttributeEffect("2x") is 2
PASS colspanAttributeEffect("2.") is 2
PASS colspanAttributeEffect("2.9") is 2
PASS colspanAttributeEffect("a") is 1
PASS colspanAttributeEffect("7") is 1
PASS colspanAttributeEffect("  7") is 7
PASS colspanAttributeEffect(arabicIndicDigitOne) is 1
PASS colspanAttributeEffect("2" + arabicIndicDigitOne) is 2
PASS colspanAttributeEffect("2147483647") is 1000
PASS colspanAttributeEffect("4294967295") is 1000
PASS colspanAttributeEffect("4294967296") is 1000
PASS successfullyParsed is true

TEST COMPLETE
